refactor(server): rename to `unixepoch_macros`
